Refactor component imports to streamline file structure
This commit updates the import paths for various components and utilities across the application, specifically moving imports from the utils and components directories to a more organized structure under the plots directory. Components such as ModalContentAoiModification, ModalContentAoiVisibility, ModalContentDownloadBarPlot, and others have been adjusted to reflect these changes. Additionally, the RecurrencePlot and its related files have been removed as they are no longer in use. This refactor enhances code maintainability and clarity in the project's structure.
