+++ src/lib/plots/bar/components/BarPlotButtonMenu.svelte
