+++ src/lib/plots/bar/components/index.ts
@@ -1,0 +1,3 @@
+export { default as BarPlot } from './BarPlot.svelte'
+export { default as BarPlotButtonMenu } from './BarPlotButtonMenu.svelte'
+export { default as BarPlotFigure } from './BarPlotFigure.svelte'
