+++ src/lib/plots/scarf/components/ScarfPlotButtonResetView.svelte
