+++ src/lib/plots/scarf/components/ScarfPlotButtonZoomIn.svelte
