+++ src/lib/plots/scarf/components/ScarfPlotButtonZoomOut.svelte
