+++ src/lib/plots/scarf/components/ScarfPlotFigure.svelte
