+++ src/lib/plots/scarf/components/ScarfPlotSelectGroup.svelte
