+++ src/lib/plots/scarf/components/ScarfPlotSelectStimulus.svelte
@@ -2,7 +2,7 @@
   import Select from '$lib/components/General/GeneralSelect/GeneralSelect.svelte'
   import { handleScarfSelectionChange } from '$lib/services/scarfSelectService'
   import type { ScarfGridType } from '$lib/type/gridType'
-  import { getStimuliOptions } from '$lib/utils/sharedPlotUtils'
+  import { getStimuliOptions } from '$lib/plots/shared/utils/sharedPlotUtils'
 
   interface Props {
     settings: ScarfGridType
