+++ src/lib/plots/shared/class/AdaptiveTimeline.ts
