+++ src/lib/plots/shared/components/index.ts
@@ -1,0 +1,1 @@
+export { default as PlotPlaceholder } from './PlotPlaceholder.svelte'
