+++ src/lib/plots/shared/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './sharedPlotUtils'
+export * from './plotSizeUtility'
