+++ src/lib/plots/shared/utils/sharedPlotUtils.ts
@@ -1,0 +1,14 @@
+import { getStimuli } from '$lib/stores/dataStore'
+
+/**
+ * Get the stimuli options for a plot
+ * @returns {Array} The stimuli options
+ */
+export function getStimuliOptions() {
+  return getStimuli().map(stimulus => {
+    return {
+      label: stimulus.displayedName,
+      value: stimulus.id.toString(),
+    }
+  })
+}
