+++ src/lib/plots/transition-matrix/components/TransitionMatrixButtonMenu.svelte
