+++ src/lib/plots/transition-matrix/components/TransitionMatrixPlotFigure.svelte
