+++ src/lib/plots/transition-matrix/components/TransitionMatrixSelectGroup.svelte
