+++ src/lib/plots/transition-matrix/components/index.ts
@@ -1,0 +1,5 @@
+export { default as TransitionMatrixPlot } from './TransitionMatrixPlot.svelte'
+export { default as TransitionMatrixPlotFigure } from './TransitionMatrixPlotFigure.svelte'
+export { default as TransitionMatrixButtonMenu } from './TransitionMatrixButtonMenu.svelte'
+export { default as TransitionMatrixSelectStimulus } from './TransitionMatrixSelectStimulus.svelte'
+export { default as TransitionMatrixSelectGroup } from './TransitionMatrixSelectGroup.svelte'
